src/uwsgi/requirements.txt
flask
uwsgi
mysql-connector-python
bcrypt
pybase64
flask-cors
pycryptodome
redis
Flask-WTF
termcolor